Loop on data-engineer module
This demonsatrates passing in a block of blocks to a generic (ish) module
We now provide defaults to managed_policies and policy documents, so
they don't need to be specified.
This has the benefit of seperating out the data from the logic/module
call, reducing the lines of code to do the same operation for all the
data engineers.
This can only be done for one account at a time as the provider block
cannot by dynamic.